create preview blog page
