+++ src/app/create-blog/page.tsx
@@ -9,6 +9,7 @@ import { create_UUID } from "@/lib/uuid";
 import { Input } from "@/components/ui/input";
 import { useSearchParams } from "next/navigation";
 import * as Dialog from "@radix-ui/react-dialog";
+import PreviewBlog from "../preview-blog/page";
 
 export default function Page() {
   const [filename, setFilename] = useState<string>();
@@ -43,25 +44,6 @@ export default function Page() {
         },
       ];
     });
-  };
-
-  const save = () => {
-    if (filename && cards.length > 0) {
-      const text = cards?.map((item) => item.value + "\n\n").join("");
-
-      const currentLocalStorage = JSON.parse(
-        localStorage.getItem("blogs") as string
-      );
-
-      const newLocalStorage = {
-        ...currentLocalStorage,
-        [filename]: { cards, filename, text },
-      };
-
-      localStorage.setItem("blogs", JSON.stringify(newLocalStorage));
-    } else {
-      // error message
-    }
   };
 
   return (
@@ -108,22 +90,9 @@ export default function Page() {
           onOpenChange={() => setModalIsOpen(!modalIsOpen)}
         >
           <Dialog.Portal>
-            <Dialog.Overlay className="fixed inset-0 bg-black-a9" />
-            <Dialog.Content className="bg-white rounded-6px shadow-xl absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-[90vw] max-w-[450px] max-h-[85vh] p-25">
-              <div className=" m-4 flex flex-col gap-4">
-                <Dialog.Title className="font-bold">Preview</Dialog.Title>
-                <div className="overflow-y-auto w-full p-5 border rounded-lg space-y-4 ">
-                  <article className="prose dark:prose-invert">
-                    <Markdown remarkPlugins={[remarkGfm]}>
-                      {cards?.map((item) => item.value + "\n\n").join("")}
-                    </Markdown>
-                  </article>
-                </div>
-                <div className="flex flex-row justify-around">
-                  <Button onClick={() => setModalIsOpen(false)}>Cancel</Button>
-                  <Button onClick={save}>Preview & Save</Button>
-                </div>
-              </div>
+            <Dialog.Overlay className="fixed inset-0 bg-black-a9 rounded-[6px]  shadow-2xl" />
+            <Dialog.Content className="overflow-scroll bg-white rounded-[6px] shadow-2xl absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-[700px] h-[700px] p-25">
+              <PreviewBlog cards={cards} setModalIsOpen={setModalIsOpen} />
             </Dialog.Content>
           </Dialog.Portal>
         </Dialog.Root>
